+++ src/main/kotlin/io/github/x1111101101/glucoseserver/record/model/exercise/extra/dataholder/BurnedCaloriesDataHolder.kt
@@ -1,5 +1,5 @@
 package io.github.x1111101101.glucoseserver.record.model.exercise.extra.dataholder
 
 interface BurnedCaloriesDataHolder {
-    val burnedCalories: Int
+    val burnedCalories: Double
 }
